add component test cases
