+++ README.md
@@ -1,14 +1,14 @@
 # orbital-grindtogether
-A web application developed for NUS Orbital 2023
+A web application developed for NUS Orbital Programme 2023.
 
-You need to have Node.js installed in your machine to run this web app.
+You can try out our project [here](https://glittery-pastelito-97d2c4.netlify.app/).
 
 Frontend is developed using the Next.js framework and created using the command
 ```
 npx create-next-app@latest
 ```
 
-Backend is developed using the Express.js framework.
+Backend is developed using Node.js with Express.js.
 Database and Authentication is provided via Firebase itself.
 
 Command to install all dependencies
