+++ frontend-client/cypress/components/likeButton.cy.tsx
