+++ frontend-client/cypress/components/navigationBar.cy.tsx
